chore(camel-k): Update Fabric8 CamelK Extension model to v1.7.0 (#3670)
+ Bump Go Module dependencies to v1.7.0
