+++ extensions/camel-k/generator-v1alpha1/cmd/generate/generate.go
@@ -80,6 +80,7 @@ func main() {
 		reflect.TypeOf(apis.URL{}):             "java.lang.String",
 		reflect.TypeOf(apis.VolatileTime{}):    "java.lang.String",
 		reflect.TypeOf(runtime.RawExtension{}): "java.util.Map<String, Object>",
+		reflect.TypeOf(v1.Template{}):          "java.util.Map<String, Object>",
 	}
 
 	json := schemagen.GenerateSchema("http://fabric8.io/camel-k/v1alpha1/CamelKSchema#", crdLists, providedPackages, manualTypeMap, packageMapping, mappingSchema, providedTypes, constraints, "io.fabric8")
